be_sword/pyproject.toml
[tool.poetry]
name = "be-sword"
version = "0.1.0"
description = ""
authors = ["Carlos Souza <carlosspy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
djangorestframework = "^3.15.2"
model-bakery = "^1.19.5"
django-filter = "^24.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pytest-django = "^4.9.0"
black = "^24.8.0"
mypy = "^1.11.2"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
